code refactoring

+++ src/main/java/com/example/et/controller/dto/AccountDto.java
@@ -3,9 +3,8 @@ package com.example.et.controller.dto;
 import com.example.et.model.core.Account;
 import com.example.et.model.core.Bank;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
-public record AccountDto(UUID id, String lastFourDigits, BigDecimal balance, Account.AccountType accountType,
+public record AccountDto(UUID id, String lastFourDigits, Float balance, Account.AccountType accountType,
                          Bank bank) {
 }
