+++ src/main/java/com/example/et/controller/dto/CardDto.java
@@ -3,8 +3,7 @@ package com.example.et.controller.dto;
 import com.example.et.model.core.Bank;
 import com.example.et.model.core.CardType;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
-public record CardDto(UUID id, CardType cardType, String lastFourDigits, UUID accountId, BigDecimal limit, Bank bank) {
+public record CardDto(UUID id, CardType cardType, String lastFourDigits, UUID accountId, Float limit, Bank bank) {
 }
