+++ src/main/java/com/example/et/controller/dto/UserDetailsDto.java
@@ -3,12 +3,10 @@ package com.example.et.controller.dto;
 import com.example.et.model.core.AppUserConfig;
 import com.example.et.model.core.LanguagePreference;
 
-import java.math.BigDecimal;
-
 public record UserDetailsDto(String email,
                              String name,
                              boolean isOnboardingComplete,
                              LanguagePreference languagePreference,
-                             BigDecimal spendLimit,
+                             Integer spendLimit,
                              AppUserConfig.Currency currency, String paymentMode) implements UserEditableProperties {
 }
