+++ src/main/java/com/example/et/model/core/AppUserConfig.java
@@ -3,7 +3,6 @@ package com.example.et.model.core;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -20,7 +19,7 @@ public class AppUserConfig extends BaseAudit {
   @Enumerated(EnumType.STRING)
   private LanguagePreference languagePreference;
 
-  private BigDecimal spendLimit;
+  private Integer spendLimit;
 
   @Enumerated(EnumType.STRING)
   private AppUserConfig.Currency currency;
